+ separate router provider component
