+++ frontend/src/libs/components/components.ts
@@ -3,3 +3,5 @@ export { Icon } from "./icon/icon";
 export { Logo } from "./logo/logo";
 
 export { Button } from './button/button';
+
+export { RouterProvider } from './router-provider/router-provider';
